chore: move suggestors into plugin
