+++ packages/core/src/index.ts
@@ -1,4 +1,9 @@
 export { generateReport } from './report';
 export { Report, Package, Suggestion, SuggestionAction } from './models';
 export * from './types';
-export { getBreadcrumb } from './package/get-breadcrumb';
+export {
+  getBreadcrumb,
+  humanFileSize,
+  getDirectorySize,
+  getLatestPackages,
+} from './package';
