+++ packages/plugin/src/suggestors/index.ts
@@ -2,3 +2,4 @@ export { nestedDependencyFreshness } from './nested-deps-freshness';
 export { notBeingAbsorbedByTopLevel } from './not-absorbed-by-top-level';
 export { packagesWithExtraArtifacts } from './packages-with-extra-artifacts';
 export { topLevelDepsFreshness } from './top-level-deps-freshness';
+export { packagesWithPinnedVersions } from './packages-with-pinned-dependencies';
